refactor calculation of end time and queryProposals
